Rename widgets/widget.dart to widgets/framework.dart
It was confusing to have both widget.dart and widgets.dart
