+++ sky/packages/sky/example/game/lib/game_demo.dart
@@ -5,7 +5,7 @@ import 'dart:math' as Math;
 import 'sprites.dart';
 
 import 'package:sky/rendering/object.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 import 'package:sky/widgets/navigator.dart';
 import 'package:vector_math/vector_math_64.dart';
