+++ sky/packages/sky/example/game/lib/sprites.dart
@@ -15,7 +15,7 @@ import 'package:sky/base/scheduler.dart' as scheduler;
 import 'package:sky/mojo/asset_bundle.dart';
 import 'package:sky/rendering/box.dart';
 import 'package:sky/rendering/object.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 import 'package:vector_math/vector_math.dart';
 
 part 'action.dart';
