+++ sky/packages/sky/example/widgets/tabs.dart
@@ -10,7 +10,7 @@ import 'package:sky/widgets/scaffold.dart';
 import 'package:sky/widgets/tabs.dart';
 import 'package:sky/widgets/theme.dart';
 import 'package:sky/widgets/tool_bar.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 class TabbedNavigatorApp extends App {
   // The index of the selected tab for each of the TabNavigators constructed below.
