+++ sky/packages/sky/lib/widgets.dart
@@ -39,4 +39,4 @@ export 'widgets/task_description.dart';
 export 'widgets/theme.dart';
 export 'widgets/tool_bar.dart';
 export 'widgets/transitions.dart';
-export 'widgets/widget.dart';
+export 'widgets/framework.dart';
