+++ sky/packages/sky/lib/widgets/checkbox.dart
@@ -7,7 +7,7 @@ import 'dart:sky' as sky;
 import 'package:sky/rendering/object.dart';
 import 'package:sky/widgets/basic.dart';
 import 'package:sky/widgets/theme.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 import 'package:sky/rendering/toggleable.dart';
 
 export 'package:sky/rendering/toggleable.dart' show ValueChanged;
