+++ sky/packages/sky/lib/widgets/drawer_item.dart
@@ -12,7 +12,7 @@ import 'package:sky/widgets/default_text_style.dart';
 import 'package:sky/widgets/icon.dart';
 import 'package:sky/widgets/ink_well.dart';
 import 'package:sky/widgets/theme.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 typedef EventDisposition OnPressedFunction();
