+++ sky/packages/sky/lib/widgets/focus.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 typedef void FocusChanged(GlobalKey key);
