+++ sky/packages/sky/lib/widgets/icon.dart
@@ -7,7 +7,7 @@ import 'dart:sky' as sky;
 import 'package:sky/mojo/asset_bundle.dart';
 import 'package:sky/widgets/basic.dart';
 import 'package:sky/widgets/theme.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 enum IconThemeColor { white, black }
