+++ sky/packages/sky/lib/widgets/icon_button.dart
@@ -7,7 +7,7 @@ import 'dart:sky' as sky;
 import 'package:sky/rendering/box.dart';
 import 'package:sky/widgets/basic.dart';
 import 'package:sky/widgets/icon.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 class IconButton extends Component {
