+++ sky/packages/sky/lib/widgets/modal_overlay.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:sky/widgets/basic.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 class ModalOverlay extends Component {
