+++ sky/packages/sky/lib/widgets/tabs.dart
@@ -20,7 +20,7 @@ import 'package:sky/widgets/icon.dart';
 import 'package:sky/widgets/ink_well.dart';
 import 'package:sky/widgets/scrollable.dart';
 import 'package:sky/widgets/theme.dart';
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 import 'package:vector_math/vector_math.dart';
 
 typedef void SelectedIndexChanged(int selectedIndex);
