+++ sky/tests/examples/card_collection.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 import '../../packages/sky/example/widgets/card_collection.dart';
 import '../resources/display_list.dart';
