+++ sky/tests/examples/styled_text.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 import '../../packages/sky/example/widgets/styled_text.dart';
 import '../resources/display_list.dart';
