+++ sky/tests/examples/tabs.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'package:sky/widgets/widget.dart';
+import 'package:sky/widgets/framework.dart';
 
 import '../../packages/sky/example/widgets/tabs.dart';
 import '../resources/display_list.dart';
